requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
dj-database-url==2.3.0
Django==5.1.7
django-minio-storage==0.5.7
djangorestframework==3.16.0
gunicorn==23.0.0
minio==7.2.15
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
python-dotenv==1.1.0
setuptools==78.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
views==0.3
wheel==0.45.1
minio
django-storages
boto3
django-storages[boto3]
django-silk>=5.0.3
